Frontend: Fix size mismatch error for compressed blobs in GetBlobOperations (#2318)
After enabling compression, we saw lots of size mismatch error log in frontend. Turns out it's because we are setting the blob size in blob properties to the size of compressed content. This PR fixes it.

Along with the fix, this PR also does:

Add more test cases for compression in GetBlobOperation
Set the default value of compression to false. We should use configuration file to change the value.