Support replication with nonblocking network client (#2324)
Start using nonblocking network client in replica thread to boost up replication speed for cross-colo replication.

This PR adds nonblocking network client support in ReplicaThread. It mimics the logic/code in OperationControllers in router package. It first divides replicas to replicate into several groups and for each group, it sends the request to the remote servers and feed the response back to the group.

This PR also adds nonblocking network client to the existing test cases., and add a new test case to test replication for multiple remote hosts at the same time.