For PutOperation select a partition that is not READ_ONLY (#2355)
When selecting a partition for PutOperation we want to select partitions that aren't READ_ONLY, as opposed to selecting partitions that are READ_WRITE. This is to ensure compatibility as we roll out changes for partitions to be in partially readable status along with READ_ONLY and READ_WRITE state. Until the change is fully rolled out, we will accept both READ_WRITE and partially writable partitions for PutOperations.