Record replication lag in seconds for the originating dc (#2377)
This PR changes the way we record replication lag in seconds metrics. Before this PR, whenever we get a blob from a remote peer, we just record the replication lag in seconds against this peer's datacenter. However, this peer might get this blob from other datacenter. This would skew the replication lag in seconds metric.

This PR fixes this issue.