Support getting bootstrap replica in Helix Full auto mode (#2413)
This PR adds changes to create a new AmbryReplica object dynamically during OFFLINE -> BOOTSTRAP instance in Full Auto mode.

In Semi-auto mode, the new replica information such as mount path of the disk, size of the replica is decided before hand and stored to Helix (/PROPERTYSTORE/AdminConfigs/ReplicaAddition Znode). But, in Full-auto mode, the decision of the disk selection has to be made locally. This PR adds these changes.