Support FULL_AUTO compatible resource creation in HelixBootstrapUpgradeTool (#2399)
This PR is the preparation for switching to FULL_AUTO in Ambry. Ambry's resources were created by appending up to 100 partitions at one resource. It's not compatible with FULL_AUTO. In order to be compatible with FULL_AUTO, the existing resources have to be reconstructed so that one ambry-server instance can only have partitions that belong to same resource.
What does that mean? It means that if we a set of partitions that are assigned to a set of ambry-server instances and this set of instances only has those partitions, then we can create a resource for these partitions and this server instances. In this way, when we turn FULL_AUTO on this resource, servers under this resource would also in FULL_AUTO mode, since these servers only have partitions from this resource.

For exmple: for a given partition layout:
{ "partition1": {"host1", "host2", "host3" }, "partition2": {"host4", "host5", "host6" }, "partition3": {"host1", "host4", "host6"}, "partition4": {"host10", "host11", "host12"}, "partition5": {"host12", "host13", "host14"}, "partition6": {"host14", "host15", "host16"} }
host1 to host6 share partition 1, 2, 3 so we can create a resource for partition 1, 2, 3 and host1 to host6.
host10 to host16 share partition 4, 5, 6 so we can create a different resource for partition 4, 5, 6, and host10 to host16.
Different resources don't share partitions and don't share hosts.
We can of course create a resource for all these partitions and hosts, and it still be compatible with FULL_AUTO, so we have to put a cap on how many hosts can be put under one resource.

This PR update HelixBootstrapUpgradeTool to start create resource in that way.