StatsManager: add a real listener to remove host from account stats store (#2442)
This PR adds a real listener in StatsManager to remove hosts from account stats store when the host is removed from clustermap. However, there are some challenges to this. All the ambry server would listen to this clustermap change event and run the same piece of code to remove hosts from account stats store. We don't need all those ambry-servers to do the same thing, so we need some mechanism to coordinate those hosts to make sure there is only one host executing this task.

In order to do so, we are using helix's distributed lock. We will create a distributed lock for each host that is removed, and all the ambry-server hosts would try to acquire this lock. However, only one server host would succeed. That's how we guarantee that there is only one ambry-server doing this work.