Allow empty replicas to be deactivated immediately since there is no need for peers to sync up (#2443)
An empty replica still has a log size of 18 bytes for its header. Due to that, during it's deactivation process, we think that its peers are 18 bytes behind and wait indefinitely for them to catch up during INACTIVE to OFFLINE transition. This PR allows empty replica to go ahead with deactivation process without checking for catch up status.