[Full-auto] Handle transition of a replica from offline to dropped state on host start up (#2454)
In Helix full-auto mode, when a host is down, it is possible that its replicas are reassigned to a different host. After the host comes up, Helix may directly transition the replicas from OFFLINE -> DROPPED state without going through OFFLINE -> BOOTSTRAP -> STANDBY -> INACTIVE -> OFFLINE -> DROPPED steps. In this case, we have to make sure the 'replica decommission' steps are followed to make sure peers have caught up with it before deleting the files.